@@ -9,25 +9,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.11",
+    "@angular/animations": "^14.0.2",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "^13.1.3",
-    "@angular/compiler": "^13.1.3",
-    "@angular/core": "^13.1.3",
-    "@angular/forms": "^13.1.3",
+    "@angular/common": "^14.0.2",
+    "@angular/compiler": "^14.0.2",
+    "@angular/core": "^14.0.2",
+    "@angular/forms": "^14.0.2",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.1.3",
-    "@angular/platform-browser-dynamic": "^13.1.3",
-    "@angular/router": "^13.1.3",
+    "@angular/platform-browser": "^14.0.2",
+    "@angular/platform-browser-dynamic": "^14.0.2",
+    "@angular/router": "^14.0.2",
     "ngx-clipboard": "^15.1.0",
     "rxjs": "~7.4.0",
     "tslib": "^2.4.0",
     "zone.js": "^0.11.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.1.4",
-    "@angular/cli": "^13.1.4",
-    "@angular/compiler-cli": "^13.1.3",
+    "@angular-devkit/build-angular": "^14.0.2",
+    "@angular/cli": "^14.0.2",
+    "@angular/compiler-cli": "^14.0.2",
     "@types/jasmine": "^3.10.6",
     "@types/node": "^12.20.55",
     "jasmine-core": "~3.10.0",
@@ -36,8 +36,8 @@
     "karma-coverage": "^2.1.1",
     "karma-jasmine": "^4.0.2",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "ng-packagr": "^13.3.1",
-    "typescript": "^4.5.5"
+    "ng-packagr": "^14.0.2",
+    "typescript": "^4.7.3"
   },
   "version": "0.0.1"
 }
